Remove Bootstrap examples
