+++ app/build.gradle
@@ -17,7 +17,7 @@ android {
     buildToolsVersion '26.0.2'
 
     defaultConfig {
-        applicationId "ar.com.wolox.android.example"
+        applicationId "ar.com.wolox.android.training"
         minSdkVersion 19
         targetSdkVersion 26
         versionCode 1
