+++ app/src/main/java/ar/com/wolox/android/LogIn/BaseConfiguration.java
@@ -1,0 +1,7 @@
+package ar.com.wolox.android.LogIn;
+
+class BaseConfiguration {
+
+    public static final String EXAMPLE_CONFIGURAITON_KEY = "com.example.test.test";
+
+}
