+++ app/src/main/java/ar/com/wolox/android/LogIn/ui/login/LogInPresenter.java
@@ -1,0 +1,14 @@
+package ar.com.wolox.android.LogIn.ui.login;
+
+
+import ar.com.wolox.wolmo.core.presenter.BasePresenter;
+
+public class LogInPresenter extends BasePresenter<LogInView> {
+
+    // Constructor
+    public LogInPresenter(LogInView viewInstance) {
+        super(viewInstance);
+    }
+
+
+}
