+++ app/src/main/java/ar/com/wolox/android/LogIn/ui/login/LogInView.java
@@ -1,0 +1,4 @@
+package ar.com.wolox.android.LogIn.ui.login;
+
+public interface LogInView {
+}
