--- app/src/main/java/ar/com/wolox/android/example/BaseConfiguration.java
@@ -1,7 +1,0 @@
-package ar.com.wolox.android.example;
-
-class BaseConfiguration {
-
-    public static final String EXAMPLE_CONFIGURAITON_KEY = "http://www.example.com";
-
-}
