--- app/src/main/java/ar/com/wolox/android/example/ui/ExampleView.java
@@ -1,8 +1,0 @@
-package ar.com.wolox.android.example.ui;
-
-public interface ExampleView {
-
-    // The example View should be able to update its UI if the random number changes
-    void onRandomNumberUpdate(int someNumber);
-
-}
