+++ app/src/main/res/values/strings.xml
@@ -1,9 +1,10 @@
 <resources>
 
     <!-- General -->
-    <string name="example_title">ExampleFragment</string>
-    <string name="example_textview">Some example TextView</string>
-    <string name="example_randomize">Randomize</string>
-    <string name="example_wolox">WOLOX</string>
-    <string name="example_message">The presenter generated the number %1$d and this view is displaying it</string>
+    <string name="login_email_address">Email address</string>
+    <string name="login_password">Password</string>
+    <string name="login_login">Log In</string>
+    <string name="login_signup">Sign Up</string>
+    <string name="login_terms_conditions">By joining you agree to our Terms &amp; Conditions</string>
+    <string name="login_message">The presenter generated the number %1$d and this view is displaying it</string>
 </resources>
