+++ app/src/production/java/ar/com/wolox/android/LogIn/Configuration.java
@@ -1,4 +1,4 @@
-package ar.com.wolox.android.example;
+package ar.com.wolox.android.LogIn;
 
 public class Configuration extends BaseConfiguration {
     public static final String API_ENDPOINT = "http://api.wolox.com.ar";
