+++ app/src/stage/java/ar/com/wolox/android/example/Configuration.java
@@ -1,4 +1,4 @@
-package ar.com.wolox.android.example;
+package ar.com.wolox.android.LogIn;
 
 public class Configuration extends BaseConfiguration {
     public static final String API_ENDPOINT = "http://stage.api.wolox.com.ar";
