+++ app/src/test/java/ar/com/wolox/android/LogIn/LogInPresenterTest.java
@@ -1,0 +1,5 @@
+package ar.com.wolox.android.LogIn;
+
+import android.content.Context;
+
+public class LogInPresenterTest {}
